# server/docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpsPilot API
  version: 0.7.0-dev
  description: OpsPilot auth, RBAC, Agent, Host, script, task execution, logs, scheduling, webhooks, metrics, alerts, and notifications API.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/v1/workflows:
    get:
      summary: List workflow definitions
      description: Requires `workflow:read`.
      parameters:
        - $ref: "#/components/parameters/keyword"
        - name: status
          in: query
          schema: { type: string, enum: [draft, active, disabled, archived] }
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Paged workflow definition list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDefinitionPage"
    post:
      summary: Create workflow definition
      description: Requires `workflow:manage`. The `definition` field is a JSON string containing a DAG with nodes and edges.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowDefinitionPayload"
      responses:
        "200":
          description: Workflow definition detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDefinitionDetail"
  /api/v1/workflows/{id}:
    get:
      summary: Get workflow definition
      description: Requires `workflow:read`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Workflow definition detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDefinitionDetail"
    put:
      summary: Update workflow definition
      description: Requires `workflow:manage`. Updating an active workflow returns it to draft status.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowDefinitionPayload"
      responses:
        "200": { description: Workflow definition detail }
  /api/v1/workflows/{id}/publish:
    post:
      summary: Publish workflow definition
      description: Requires `workflow:manage`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200": { description: Workflow definition detail }
  /api/v1/workflows/{id}/disable:
    post:
      summary: Disable workflow definition
      description: Requires `workflow:manage`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200": { description: Workflow definition detail }
  /api/v1/workflows/{id}/copy:
    post:
      summary: Copy workflow definition
      description: Requires `workflow:manage`.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        "200": { description: Workflow definition detail }
  /api/v1/workflows/{id}/versions:
    get:
      summary: List workflow versions
      description: Requires `workflow:read`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Workflow versions
          content:
            application/json:
              schema:
                type: object
                properties:
                  code: { type: integer }
                  message: { type: string }
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/WorkflowVersionSummary"
  /api/v1/workflows/{id}/run:
    post:
      summary: Start workflow run
      description: Requires `workflow:execute`. Creates a workflow run, node records, and queues root task nodes through the existing Task Run executor.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowRunPayload"
      responses:
        "200":
          description: Workflow run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunDetail"
  /api/v1/workflow-runs:
    get:
      summary: List workflow runs
      description: Requires `workflow:read`.
      parameters:
        - $ref: "#/components/parameters/keyword"
        - name: status
          in: query
          schema: { type: string, enum: [pending, queued, running, canceling, success, failed, canceled] }
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Paged workflow run list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunPage"
  /api/v1/workflow-runs/{id}:
    get:
      summary: Get workflow run detail
      description: Requires `workflow:read`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Workflow run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowRunDetail"
  /api/v1/workflow-runs/{id}/cancel:
    post:
      summary: Cancel workflow run
      description: Requires `workflow:manage`.
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200": { description: Workflow run detail }
  /api/v1/workflow-runs/{id}/retry:
    post:
      summary: Retry workflow run
      description: Requires `workflow:execute`.
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200": { description: Workflow run detail }
  /api/v1/workflow-runs/{id}/nodes/{nodeId}/approve:
    post:
      summary: Approve workflow approval node
      description: Requires `workflow:execute`.
      parameters:
        - $ref: "#/components/parameters/id"
        - name: nodeId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment: { type: string }
      responses:
        "200": { description: Workflow run detail }
  /api/v1/workflow-runs/{id}/nodes/{nodeId}/reject:
    post:
      summary: Reject workflow approval node
      description: Requires `workflow:execute`.
      parameters:
        - $ref: "#/components/parameters/id"
        - name: nodeId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comment: { type: string }
      responses:
        "200": { description: Workflow run detail }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentTokenAuth:
      type: http
      scheme: bearer
      description: Agent Token, not user JWT.
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
    keyword:
      name: keyword
      in: query
      schema: { type: string }
...
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    pageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  schemas:
...
    WorkflowDefinitionPayload:
      type: object
      required: [name, definition]
      properties:
        name: { type: string }
        description: { type: string }
        definition:
          type: string
          description: JSON string containing nodes, edges, maxParallel, and failurePolicy.
          example: '{"nodes":[{"id":"collect","type":"task","name":"Collect diagnostics"}],"edges":[],"maxParallel":1,"failurePolicy":"stop_on_failure"}'
    WorkflowDefinitionSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        version: { type: integer }
        status: { type: string, enum: [draft, active, disabled, archived] }
        nodeCount: { type: integer }
        edgeCount: { type: integer }
        createdBy: { type: string }
        publishedAt: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    WorkflowDefinitionDetail:
      allOf:
        - $ref: "#/components/schemas/WorkflowDefinitionSummary"
        - type: object
          properties:
            definition: { type: string, description: Normalized workflow DAG JSON string. }
    WorkflowDefinitionPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowDefinitionSummary"
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
    WorkflowVersionSummary:
      type: object
      properties:
        id: { type: string }
        workflowId: { type: string }
        version: { type: integer }
        status: { type: string, enum: [draft, published, deprecated, archived] }
        definitionHash: { type: string }
        createdBy: { type: string }
        publishedAt: { type: string }
        createdAt: { type: string }
    WorkflowRunPayload:
      type: object
      properties:
        triggerType: { type: string, enum: [manual, schedule, webhook, incident, api, retry], default: manual }
        idempotencyKey: { type: string }
        input: { type: string, description: JSON string input for the workflow run. }
    WorkflowRunSummary:
      type: object
      properties:
        id: { type: string }
        workflowId: { type: string }
        workflowName: { type: string }
        workflowVersion: { type: integer }
        status: { type: string, enum: [pending, queued, running, canceling, success, failed, canceled] }
        triggerType: { type: string }
        totalNodes: { type: integer }
        successNodes: { type: integer }
        failedNodes: { type: integer }
        skippedNodes: { type: integer }
        errorMessage: { type: string }
        createdBy: { type: string }
        queuedAt: { type: string }
        startedAt: { type: string }
        finishedAt: { type: string }
        createdAt: { type: string }
    WorkflowRunNode:
      type: object
      properties:
        id: { type: integer }
        nodeId: { type: string }
        nodeType: { type: string, enum: [task, condition, approval, notification, webhook, wait, incident] }
        nodeName: { type: string }
        status: { type: string, enum: [pending, queued, running, success, failed, skipped, canceled] }
        taskRunId: { type: string }
        errorMessage: { type: string }
        attempts: { type: integer }
        queuedAt: { type: string }
        startedAt: { type: string }
        finishedAt: { type: string }
    WorkflowRunEvent:
      type: object
      properties:
        id: { type: integer }
        nodeId: { type: string }
        eventType: { type: string }
        message: { type: string }
        actor: { type: string }
        payload: { type: string, description: JSON string payload. }
        createdAt: { type: string }
    WorkflowRunDetail:
      allOf:
        - $ref: "#/components/schemas/WorkflowRunSummary"
        - type: object
          properties:
            input: { type: string }
            output: { type: string }
            definition: { type: string, description: Workflow definition snapshot JSON string. }
            nodes:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowRunNode"
            events:
              type: array
              items:
                $ref: "#/components/schemas/WorkflowRunEvent"
    WorkflowRunPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowRunSummary"
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...